MODIFY : comment App 삭제
- 기존 comment app에 있던 model, view, url은 movie로 이동
