--- comment/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class CommentConfig(AppConfig):
-    name = 'comment'
